chore: Rename time module to std_time.
